@@ -6,27 +6,39 @@
       "@reduxjs/toolkit": "^1.6.1",
       "@types/lodash": "^4.14.172",
       "@types/node": "^12.0.0",
+<<<<<<< HEAD
       "@types/qs": "^6.9.7",
       "@types/react": "^17.0.19",
       "@types/react-dom": "^17.0.9",
+=======
+      "@types/react": "^17.0.0",
+      "@types/react-bootstrap": "^0.32.28",
+      "@types/react-dom": "^17.0.0",
+>>>>>>> 038f15c69d6b03df4013fed46c3b2be12d2abd0b
       "@types/react-select": "^4.0.17",
       "aphrodite": "^2.4.0",
       "axios": "^0.21.1",
+      "bootstrap": "5.1.3",
       "classnames": "^2.3.1",
+      "history": "^5.1.0",
       "http-proxy-middleware": "^2.0.1",
       "intersection-observer": "^0.12.0",
-      "qs": "^6.10.1",
-      "react": "^16.9.0",
+      "react": "^17.0.2",
+      "react-bootstrap": "^2.0.2",
       "react-cool-img": "^1.2.12",
-      "react-dom": "^16.14.0",
+      "react-dom": "^17.0.2",
       "react-icons": "^4.2.0",
       "react-redux": "^7.2.4",
-      "react-router-dom": "^5.2.0",
+      "react-router": "^6.0.2",
+      "react-router-dom": "^6.0.2",
       "react-scripts": "4.0.3",
-      "react-select": "^4.3.1",
       "react-use": "^17.2.4",
       "sass": "^1.32.12",
+<<<<<<< HEAD
       "typescript": "^4.2.0"
+=======
+      "typescript": "^4.1.2"
+>>>>>>> 038f15c69d6b03df4013fed46c3b2be12d2abd0b
    },
    "scripts": {
       "start": "react-scripts start",
@@ -36,12 +48,11 @@
    },
    "eslintConfig": {
       "extends": [
-         "react-app",
-         "react-app/jest"
+         "react-app"
       ],
       "rules": {
-         "@typescript-eslint/no-unused-vars": 0,
-         "react-hooks/exhaustive-deps": 0
+         "no-unused-vars": 0,
+         "@typescript-eslint/no-unused-vars": 0
       }
    },
    "browserslist": {
@@ -51,9 +62,9 @@
          "not op_mini all"
       ],
       "development": [
-         "last 1 chrome version",
-         "last 1 firefox version",
-         "last 1 safari version"
+         "last 2 chrome version",
+         "last 2 firefox version",
+         "last 2 safari version"
       ]
    }
 }
