@@ -1,16 +1,26 @@
-import { Link } from 'react-router-dom'
-import styles from '../../style.module.scss'
+import { Button } from 'react-bootstrap';
+import img from '../../img/temp2.png';
 
 const NotFound = () => (
-  <div className={styles.NotFoundBlock}>
-    <h1>404 Not Found</h1>
-    <code className={styles.NotFoundCodeBlock}>
-      The Page you&apos;re looking for isn&apos;t here :P.
-    </code>
-    <Link className={styles.GoHomeLink} to="/">
-      Go Home Now
-    </Link>
-  </div>
-)
+   <div
+      className="position-fixed top-0 start-0 container-fluid"
+      style={{
+         height: '100vh',
+         display: 'grid',
+         placeItems: 'center',
+         alignContent: 'center',
+      }}
+   >
+      <h1 className="text-danger">404 Not Found</h1>
+      <img src={img} alt="kono dio da" height={window.innerHeight / 2} />
+      <Button
+         variant="info"
+         className="mt-2"
+         onClick={() => window.history.back()}
+      >
+         Go Home Now
+      </Button>
+   </div>
+);
 
-export default NotFound
+export default NotFound;
