@@ -1,15 +1,14 @@
 import { FC } from 'react';
 import Image from 'react-cool-img';
-import errorImg from '../img/404.jpg';
-import { MangaViewProps } from './props';
+import { MangaViewProps } from '../props';
 
 const MangaView: FC<MangaViewProps> = ({ panelImg }) => (
-   <div className="panel">
+   <div>
       <Image
          src={panelImg.path}
          alt={panelImg.name}
-         error={errorImg}
-         style={!panelImg ? { background: '#444', height: '500px' } : {}}
+         error={'/kono_dio_da.png'}
+         style={{ width: '100%' }}
       />
    </div>
 );
