@@ -15,10 +15,12 @@ export interface MangaData {
    path: string;
 }
 
-export type SearchResult = Fuse.FuseResult<Manga>;
+export type SearchResult = {
+   item: Manga;
+};
 
 export type MangasResult = {
    items: Manga[];
    first: number;
    last: number;
-} & Record<'next' | 'prev' | 'total' | 'limit', number>;
+} & Record<'next' | 'prev' | 'total' | 'limit' | 'current', number>;
